refactor: Remove barrel index, update imports to direct paths
