+++ src/components/Breadcrumbs/Breadcrumbs.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import React from 'react';
-import { Link } from '../Link/Link';
+import { Link } from '../Link';
 import {
   BreadcrumbItem,
   BreadcrumbLink,
