+++ src/components/Checkbox/Checkbox.stories.tsx
@@ -1,6 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { type ChangeEvent, useEffect, useRef, useState } from 'react';
-import { ErrorText, Legend, RequirementBadge, SupportText } from '../';
+import { ErrorText } from '../ErrorText';
+import { Legend } from '../Legend';
+import { RequirementBadge } from '../RequirementBadge';
+import { SupportText } from '../SupportText';
 import { Checkbox } from './Checkbox';
 
 const meta = {
