+++ src/components/DatePicker/DatePicker.stories.tsx
@@ -13,19 +13,19 @@ import {
   Popover,
   DatePicker as ReactAriaDatePicker,
 } from 'react-aria-components';
+import { Button } from '../Button';
 import {
-  Button,
   DatePicker,
   DatePickerCalendarButton,
   DatePickerDate,
   DatePickerMonth,
   DatePickerYear,
-  ErrorText,
-  Legend,
-  RequirementBadge,
-  Select,
-  SupportText,
-} from '..';
+} from '../DatePicker';
+import { ErrorText } from '../ErrorText';
+import { Legend } from '../Legend';
+import { RequirementBadge } from '../RequirementBadge';
+import { Select } from '../Select';
+import { SupportText } from '../SupportText';
 
 /**
  * Defaultタイプのデートピッカー。
