+++ src/components/Disclosure/Disclosure.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { List } from '../';
+import { List } from '../List';
 import { Disclosure, DisclosureBackLink, DisclosureSummary } from './Disclosure';
 
 const meta = {
