+++ src/components/NotificationBanner/parts/Body.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { Button } from '../../';
+import { Button } from '../../Button';
 import { NotificationBannerBody } from './Body';
 
 const meta = {
