+++ src/components/Select/Select.stories.tsx
@@ -1,6 +1,10 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import React from 'react';
-import { ErrorText, Label, Legend, RequirementBadge, SupportText } from '../';
+import { ErrorText } from '../ErrorText';
+import { Label } from '../Label';
+import { Legend } from '../Legend';
+import { RequirementBadge } from '../RequirementBadge';
+import { SupportText } from '../SupportText';
 import { Select } from './Select';
 
 const meta = {
