--- src/components/index.ts
@@ -1,83 +1,0 @@
-export {
-  Accordion,
-  AccordionBackLink,
-  AccordionContent,
-  AccordionSummary,
-} from './Accordion';
-export { Blockquote } from './Blockquote';
-export { BreadcrumbItem, BreadcrumbList, Breadcrumbs, BreadcrumbsLabel } from './Breadcrumbs';
-export { Button, buttonBaseStyle, buttonSizeStyle, buttonVariantStyle } from './Button';
-export { Carousel, CarouselSingle, CarouselSingleImage, CarouselSingleLink } from './Carousel';
-export { Checkbox } from './Checkbox';
-export { ChipLabel } from './ChipLabel';
-export * from './DatePicker';
-export { Disclosure, DisclosureBackLink, DisclosureSummary } from './Disclosure';
-export { Divider } from './Divider';
-export { Dd, Dl, Dt } from './Dl';
-export {
-  EmergencyBanner,
-  EmergencyBannerBody,
-  EmergencyBannerButton,
-  EmergencyBannerHeading,
-} from './EmergencyBanner';
-export { ErrorText } from './ErrorText';
-export {
-  FileUpload,
-  FileUploadDropArea,
-  FileUploadFileInfo,
-  FileUploadFileItem,
-  FileUploadFileList,
-  FileUploadFileMarker,
-  FileUploadFileMeta,
-  FileUploadFileName,
-  FileUploadInput,
-  FileUploadViewportOverlay,
-  FileUploadViewportOverlayMessage,
-  fileUploadDefaultMessages,
-} from './FileUpload';
-export {
-  CloseIcon,
-  CloseWithLabelIcon,
-  HamburgerIcon,
-  HamburgerMenuButton,
-  HamburgerWithLabelIcon,
-} from './HamburgerMenuButton';
-export { Heading, HeadingShoulder, HeadingTitle } from './Heading';
-export { Input } from './Input';
-export { Label } from './Label';
-export {
-  LanguageSelector,
-  LanguageSelectorArrowIcon,
-  LanguageSelectorButton,
-  LanguageSelectorGlobeIcon,
-  LanguageSelectorGlobeWithLabelIcon,
-  LanguageSelectorMenu,
-  LanguageSelectorMenuItem,
-} from './LanguageSelector';
-export { Legend } from './Legend';
-export {
-  Link,
-  LinkExternalLinkIcon,
-  linkActiveStyle,
-  linkDefaultStyle,
-  linkFocusStyle,
-  linkHoverStyle,
-  linkStyle,
-  linkVisitedStyle,
-} from './Link';
-export { List, listBaseStyle, listDefaultStyle, listNumberedStyle } from './List';
-export {
-  NotificationBanner,
-  NotificationBannerBody,
-  NotificationBannerClose,
-  NotificationBannerMobileClose,
-} from './NotificationBanner';
-export { Radio } from './Radio';
-export { RequirementBadge } from './RequirementBadge';
-export { Select } from './Select';
-export * from './SeparatedDatePicker';
-export { Slot } from './Slot';
-export { StatusBadge } from './StatusBadge';
-export { SupportText } from './SupportText';
-export { Textarea } from './Textarea';
-export { UtilityLink, UtilityLinkExternalLinkIcon, utilityLinkStyle } from './UtilityLink';
